+++ modules/tinymce/src/core/main/ts/content/GetContent.ts
@@ -5,15 +5,13 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Cell } from '@ephox/katamari';
-
+import { GetContentRegistry } from '../api/content/EditorContent';
 import Editor from '../api/Editor';
 import * as Rtc from '../Rtc';
 import { Content, GetContentArgs } from './ContentTypes';
-import { GetContentFormatter } from './EditorContent';
 
 const defaultFormat = 'html';
-export const getContent = (editor: Editor, args: Partial<GetContentArgs> = {}, getCell: Cell<Record<string, GetContentFormatter>>): Content => {
+export const getContent = (editor: Editor, args: Partial<GetContentArgs> = {}, getCell: GetContentRegistry): Content => {
   const format = args.format ? args.format : defaultFormat;
 
   return Rtc.getContent(editor, args, format, getCell);