+++ modules/tinymce/src/core/main/ts/content/RegisterFormat.ts
@@ -1,4 +1,3 @@
-import { Cell } from '@ephox/katamari';
 /**
  * Copyright (c) Tiny Technologies, Inc. All rights reserved.
  * Licensed under the LGPL or a commercial license.
@@ -6,9 +5,9 @@ import { Cell } from '@ephox/katamari';
  * For commercial licenses see https://www.tiny.cloud/
  */
 
+import { GetContentCallback, GetContentRegistry, SetContentCallback, SetContentRegistry } from '../api/content/EditorContent';
 import Editor from '../api/Editor';
 import * as Rtc from '../Rtc';
-import { GetContentFormatter, SetContentFormatter } from './EditorContent';
 
-export const addFormat = (editor: Editor, format: string, formatGetter: GetContentFormatter, getCell: Cell<Record<string, GetContentFormatter>>, formatSetter: SetContentFormatter, setCell: Cell<Record<string, SetContentFormatter>>): void =>
+export const registerFormat = (editor: Editor, format: string, formatGetter: GetContentCallback, getCell: GetContentRegistry, formatSetter: SetContentCallback, setCell: SetContentRegistry): void =>
   Rtc.addContentFormat(editor, format, formatGetter, getCell, formatSetter, setCell);